+++ react-app/src/App.js
@@ -1,71 +1,80 @@
-import React, { useState } from 'react';
+// Imports
+
+import React, {useState} from 'react';
 import Login from './Login';
 import ChatList from './ChatList';
 import Chat from './Chat';
 import NewChat from './NewChat';
 
+// Main application
+
 const App = () => {
-  const [loggedIn, setLoggedIn] = useState(false);
-  const [selectedChatId, setSelectedChatId] = useState(null);
-  const [showNewChat, setShowNewChat] = useState(false);
-  const [chats, setChats] = useState([]);
 
-  const handleLogin = (username, password) => {
-    // Perform login logic here
-    setLoggedIn(true);
-  };
+    // Variable Declarations
 
-  const handleLogout = () => {
-    setLoggedIn(false);
-  };
+    const [loggedIn, setLoggedIn] = useState(false); // Boolean for logging in
+    const [selectedChatId, setSelectedChatId] = useState(null); // Integer for chat selection
+    const [chats, setChats] = useState([]); // List of chats
 
-  const handleSelectChat = (chatId) => {
-    setSelectedChatId(chatId);
-  };
+    // Handler for logging in, pending login logic
 
-  const handleNewChat = () => {
-    setShowNewChat(true);
-  };
+    const handleLogin = (username, password) => {
+        setLoggedIn(true);
+    };
 
-  const handleCreateChat = (users, name) => {
-    const newChatId = Math.max(...chats.map((chat) => chat.id)) + 1;
-    const newChat = { id: newChatId, users, name, messages: [] };
-    setChats([...chats, newChat]);
-    setSelectedChatId(newChatId);
-  };
+    // Handler for logging out
 
-  const handleNewChatClose = () => {
-    setShowNewChat(false);
-  };
+    const handleLogout = () => {
+        setLoggedIn(false);
+    };
 
-  if (!loggedIn) {
-    return <Login onLogin={handleLogin} />;
-  }
+    // Handler for clicking on a chat to view it
 
-  if (selectedChatId) {
-    const selectedChat = chats.find((chat) => chat.id === selectedChatId);
-    return <Chat chat={selectedChat} setSelectedChatId={setSelectedChatId} />;
-  }
+    const handleSelectChat = (chatId) => {
+        setSelectedChatId(chatId);
+    };
 
-  return (
-    <div className='page'>
-      <h1>CrewConnect</h1>
-      <NewChat
-        show={showNewChat}
-        onClose={handleNewChatClose}
-        onCreate={handleCreateChat}
-      />
-      <ChatList
-        chats={chats}
-        onSelect={handleSelectChat}
-        onNewChat={handleNewChat}
-      />
-      {selectedChatId && (
-        <Chat chat={chats.find((chat) => chat.id === selectedChatId)} setSelectedChatId={setSelectedChatId} />
-      )}
-      <button onClick={handleLogout}>Log Out</button>
-    </div>
-  );
+    // Handler for creating a new chat
+
+    const handleCreateChat = (users, name) => {
+        const newChatId = Math.max(...chats.map((chat) => chat.id)) + 1;
+        const newChat = { id: newChatId, users, name, messages: [] };
+        setChats([...chats, newChat]);
+        setSelectedChatId(newChatId);
+    };
+
+    // Showing the login page
+
+    if (!loggedIn) {
+        return <Login onLogin={handleLogin} />;
+    }
+
+    // Showing the selected chat
+
+    if (selectedChatId) {
+        const selectedChat = chats.find((chat) => chat.id === selectedChatId);
+        return <Chat chat={selectedChat} setSelectedChatId={setSelectedChatId} />;
+    }
+
+    // Return main HTML for home page
+
+    return (
+        <div className='page'>
+            <h1>CrewConnect</h1>
+            <NewChat
+                show={selectedChatId != null}
+                onCreate={handleCreateChat}
+            />
+            <ChatList
+                chats={chats}
+                onSelect={handleSelectChat}
+            />
+            {selectedChatId && (
+                <Chat chat={chats.find((chat) => chat.id === selectedChatId)} setSelectedChatId={setSelectedChatId} />
+            )}
+            <button onClick={handleLogout}>Log Out</button>
+        </div>
+    );
   
 };
