+++ react-app/src/ChatList.js
@@ -1,19 +1,23 @@
+// Imports
+
 import React from 'react';
 
-const ChatList = ({ chats, onSelect, onNewChat }) => {
+// Just fetching the list of chats
 
-return (
-    <div className='page'>
-        <h4>Your Chats</h4>
-        <ul>
-            {chats.map((chat) => (
-                <button key={chat.id} onClick={() => onSelect(chat.id)}>
-                    {chat.name}
-                </button>
-            ))}
-        </ul>
-    </div>
-);
+const ChatList = ({chats, onSelect}) => {
+
+    // Return chat list HTML
+
+    return (
+        <div className='page'>
+            <h4>Your Chats</h4>
+                {chats.map((chat) => (
+                    <button key={chat.id} onClick={() => onSelect(chat.id)}>
+                        {chat.name}
+                    </button>
+                ))}
+        </div>
+    );
 };
 
 export default ChatList;
