+++ react-app/src/index.css
@@ -5,9 +5,142 @@ body {
     sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
+  background-color: #313338;
 }
 
 code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
+}
+
+.container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  height: 100vh;
+  background-color: #313338;
+}
+
+.container h1 {
+  color: white;
+}
+
+.container p {
+  color: #8a8e94;
+}
+
+.form {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 100px;
+}
+
+.form input {
+  margin-bottom: 10px;
+  padding: 10px;
+  border-radius: 5px;
+  border: 1px solid #ccc;
+}
+
+.form button {
+  padding: 10px;
+  border-radius: 5px;
+  border: none;
+  background-color: #7e46f4;
+  color: #fff;
+  font-size: 16px;
+  cursor: pointer;
+}
+
+.form button:hover {
+  background-color: #8962ee;
+}
+
+.page {
+  padding: 10px;
+  background-color: #313338;
+  color: #d0d0d0;
+}
+
+.page button {
+  padding: 10px;
+  border-radius: 5px;
+  margin-left: 10px;
+  border: none;
+  background-color: #7e46f4;
+  color: #fff;
+  font-size: 16px;
+  cursor: pointer;
+}
+
+.page ul {
+  list-style-type: none;
+  margin: 0;
+  padding: 0;
+  margin-bottom: 5px;
+}
+
+/* Styles for pop-ups */
+
+.popup-box {
+  position: fixed;
+  background: #00000050;
+  width: 100%;
+  height: 100vh;
+  top: 0;
+  left: 0;
+}
+ 
+.box {
+  position: relative;
+  width: 50%;
+  margin: 0 auto;
+  height: auto;
+  max-height: 70vh;
+  margin-top: calc(100vh - 85vh - 20px);
+  background: #313338;
+  border-radius: 4px;
+  padding: 20px;
+  overflow: auto;
+  border: 1px solid #999;
+  border-radius: 5px;
+  box-shadow: 0px 0px 10px #999;
+}
+ 
+.close-icon {
+  content: 'x';
+  cursor: pointer;
+  position: fixed;
+  right: 23%;
+  top: calc(100vh - 85vh - 33px);
+  background: #313338;
+  width: 25px;
+  height: 25px;    
+  border-radius: 50%;
+  line-height: 20px;
+  text-align: center;
+  border: 1px solid #999;
+  color: red;
+  font-size: 20px;
+}
+
+.popup-form button {
+  padding: 10px;
+  border-radius: 5px;
+  margin-left: 10px;
+  border: none;
+  background-color: #7e46f4;
+  color: #fff;
+  font-size: 16px;
+  cursor: pointer;
+}
+
+.popup-form input {
+  margin-bottom: 10px;
+  padding: 10px;
+  border-radius: 5px;
+  border: 1px solid #ccc;
+  margin-right: 20px;
 }
